Fix endian problem for tracepoint enabled flag
When running big endian machines there is a problem with
the enabled flag for tracepoints as it is defined as a
int8_t but written from gdbserver as an integer and then
read in the agent as 8-bit value.
This caused problem when tracepoint was disabled and
re-enabled.

gdb/gdbserver/ChangeLog:

	* tracepoint.c (struct tracepoint): Make sure to write
	enable flag as 8-bit value.

Signed-off-by: Par Olsson <par.olsson@windriver.com>
Signed-off-by: Henrik Wallin <henrik.wallin@windriver.com>